requirements.txt
argon2-cffi==20.1.0
asgiref==3.3.1
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
bleach==3.2.1
brotlipy==0.7.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.4
conda==4.8.4
conda-package-handling==1.7.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
Django==3.1.3
django-filter==2.4.0
djangorestframework==3.12.2
entrypoints==0.3
h5py==2.10.0
importlib-metadata==2.0.0
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
Jinja2==2.11.2
joblib==0.17.0
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
kiwisolver==1.3.1
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.2
menuinst==1.4.16
mistune==0.8.4
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.2
notebook==6.1.4
olefile==0.46
packaging==20.4
pandas==1.1.4
pandocfilters==1.4.3
parso==0.7.1
pickleshare==0.7.5
prometheus-client==0.8.0
prompt-toolkit==3.0.8
pycollada==0.6
pycosat==0.6.3
Pygments==2.7.2
pyparsing==2.4.7
pyreadline==2.1
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.4
pywin32==227
pywinpty==0.5.7
pyzmq==19.0.2
ruamel-yaml==0.15.87
scikit-learn==0.23.2
scipy==1.5.3
seaborn==0.11.0
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.1
traitlets==5.0.5
wcwidth==0.2.5
webencodings==0.5.1
win-inet-pton==1.1.0
wincertstore==0.2
zipp==3.4.0
